Relative import changes.

+++ landmarks_utils.py
@@ -6,7 +6,7 @@ import matplotlib.pyplot as plt
 from matplotlib.patches import Circle, Ellipse, Rectangle
 import urllib.request as urlreq
 import torch
-from config import *
+from facelandmarks.config import *
 
 
 def readtps(input, path):
@@ -226,10 +226,10 @@ def rotate_landmarks(angle_deg, landmarks, image_shape):
     rotation_matrix = cv2.getRotationMatrix2D((center[0].item(), center[1].item()), angle_deg, 1.0)
 
     centered_landmarks = pixel_landmarks - center
-    centered_landmarks = torch.matmul(centered_landmarks, torch.tensor(rotation_matrix[:,:2], dtype=torch.float32, device=DEVICE).T)
+    centered_landmarks = torch.matmul(centered_landmarks.type(torch.float), torch.tensor(rotation_matrix[:,:2], dtype=torch.float32, device=DEVICE).T)
 
     rotated_landmarks = centered_landmarks + center
-    rotated_landmarks = torch.div(rotated_landmarks, torch.tensor([image_shape[1], image_shape[0]], device=DEVICE))
+    rotated_landmarks = torch.div(rotated_landmarks.type(torch.float), torch.tensor([image_shape[1], image_shape[0]], device=DEVICE))
     return rotated_landmarks
 
 def display_parent_landmarks(projection_mask, img_idx, landmark_idx, from_inputs = True):
